file-mode-api: refactor classes to independent files
