+++ airbyte_cdk/sources/declarative/retrievers/file_uploader/file_writer.py
@@ -1,0 +1,19 @@
+#
+# Copyright (c) 2025 Airbyte, Inc., all rights reserved.
+#
+
+from pathlib import Path
+
+from airbyte_cdk.sources.declarative.retrievers.file_uploader import BaseFileWriter
+
+
+class FileWriter(BaseFileWriter):
+
+    def write(self, file_path: Path, content: bytes) -> int:
+        """
+        Writes the file to the specified location
+        """
+        with open(str(file_path), "wb") as f:
+            f.write(content)
+
+        return file_path.stat().st_size
